backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-chroma>=0.1.0
chromadb>=0.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
httpx>=0.26.0
tiktoken>=0.5.0
pypdf>=3.17.0
docx2txt>=0.8
aiofiles>=23.2.1
